# SLE-900: Speed up indexing by excluding unrelated files