… org.sonarlint.eclipse.buildship/plugin.xml
… org.sonarlint.eclipse.core/META-INF/MANIFEST.MF
… org.sonarlint.eclipse.core/plugin.xml
… org.sonarlint.eclipse.core/src/org/sonarlint/eclipse/core/internal/cache/AbstractConfigScopeIdCache.java
… org.sonarlint.eclipse.core/src/org/sonarlint/eclipse/core/internal/cache/DefaultSonarLintProjectAdapterCache.java
… org.sonarlint.eclipse.core/src/org/sonarlint/eclipse/core/internal/cache/IProjectScopeProviderCache.java
… org.sonarlint.eclipse.core/src/org/sonarlint/eclipse/core/internal/extension/SonarLintExtensionTracker.java
… org.sonarlint.eclipse.core/src/org/sonarlint/eclipse/core/internal/resources/DefaultSonarLintProjectAdapter.java
… org.sonarlint.eclipse.core/src/org/sonarlint/eclipse/core/internal/utils/SonarLintUtils.java
… org.sonarlint.eclipse.core/src/org/sonarlint/eclipse/core/resource/IProjectScopeProvider.java
… org.sonarlint.eclipse.jdt/plugin.xml
… org.sonarlint.eclipse.jdt/src/org/sonarlint/eclipse/jdt/internal/JavaProjectConfiguratorExtension.java
… org.sonarlint.eclipse.jdt/src/org/sonarlint/eclipse/jdt/internal/JdtUtils.java
… org.sonarlint.eclipse.m2e/plugin.xml
… org.sonarlint.eclipse.m2e/src/org/sonarlint/eclipse/m2e/internal/MavenModuleFilter.java
… org.sonarlint.eclipse.m2e/src/org/sonarlint/eclipse/m2e/internal/MavenUtils.java
… org.sonarlint.eclipse.pydev/plugin.xml
… org.sonarlint.eclipse.ui/src/org/sonarlint/eclipse/ui/internal/ConfigScopeIdCacheCleaner.java
… org.sonarlint.eclipse.ui/src/org/sonarlint/eclipse/ui/internal/SonarLintUiPlugin.java