## Summary

Currently, when a project is imported or a workspace is opened, all its files are indexed (with very light filtering) and sent to SLCORE for housekeeping (`SonarLintEclipseHeadlessRpcClient#listFiles()` and `FileSystemSynchronizer#getSonarLintJsonFiles()`).
When one or more projects are analyzed, this is done as well per project to get all relevant files to be sent to SLCORE for analysis (`SelectionUtils#collectFiles()`).

When any file is removed, changed, or added we listen for that and calculate that delta on every file in the project no matter whether it is relevant or not (`FileSystemSynchronizer#resourceChanged()` that invokes `FileSystemSynchronizer#visitDeltaPostChange()`).

This is very slow and inefficient, both in the performance and the high memory consumption it carries. This is especially the case for hierarchical projects as in Eclipse parent projects will contain all the sub-modules/-projects content as well but keep it hidden - even if the sub-modules/-projects are also present in the workspace.

### Compilation output (e.g. bytecode)

We want to narrow down the focus of files that are "present" and available for SonarLint even though they are not relevant. For example, the compiled code files themselves: For the Java analysis, the bytecode is necessary to yield better results, but we don't need to have all the files indexed; we just need SonarLint to know where the bytecode is stored to populate the analysis properties correctly - the bytecode cannot be analyzed itself.

For this, the JDT / Maven and Gradle sub-plugins are enhanced with a new extension point to aggregate the "output" directories, which are then excluded from the indexing.

### Sub-modules/-projects content in the parent project

We also want to narrow down the focus for projects having multiple sub-modules/-projects themselves as their files are indexed for them and shouldn't be present as well and taken into account in the parent project.

For this, the Maven and Gradle sub-plugins are enhanced by excluding all the files that are not actually part of their project scope.

### Additional, unrelated content

Additionally to this, we also want to exclude files that might be used by a build tool or version control system. In a Git repository, we don't want to index the `.git` folder as an example for a VCS. Python, on the other hand, brings libraries and tools to have a virtual Python environment directly saved in the project that we shouldn't analyze.

For this, the Python sub-plugin is enhanced by excluding common locations in a project where virtual Python environments are saved.
For version control systems we check for the most common systems like Git, [Mercurial](https://www.mercurial-scm.org/), and [Subversion](https://subversion.apache.org/) and ignore their "special" directories. This is done in `SonarLintUtils#insideVCSFolder(...)`.
For Node.js-related files, there is no sub-plugin as compared to the other sub-plugins and their changes; there is no linkable Eclipse plug-in (there is [WTP](https://eclipse.dev/webtools/), but having another optional dependency for this would be overkill). This is done in `SonarLintUtisl#inNodeJsRelated(...)`.

## Testing

Exemplary results based on hierarchical Maven projects. The check is with the `SonarLintEclipseRpcClient#listFiles()` method called when a project is indexed. This is both for speed and number of files that are indexed and stored in memory.

### [SonarLint CORE](https://github.com/SonarSource/sonarlint-core) w. 45.700  / 28 modules

When the project is clean (`git clean -dfx`), it is first imported (1), and then the workspace is re-opened (2).

| | Files before | Files after | Time before | Time after |
| --- | --- | --- | --- | --- |
| 1 | 519 | 519 | 494 | 465 |
| 2 | 1660 | 1660 | 6701 | 4088 |

When the project is dirty (`mvn clean verify -DskipTests`), it is first imported (1), and then the workspace is re-opened (2).

| | Files before | Files after | Time before | Time after |
| --- | --- | --- | --- | --- |
| 1 | 7052 | 3781 | 1277 | 465 |
| 2 | 1663 | 1660 | 9881 | 904 |

### [Orchestrator](https://github.com/SonarSource/orchestrator) w. 6.800 LOC / 4 modules

When the project is clean (`git clean -dfx`), it is first imported (1), and then the workspace is re-opened (2).

| | Files before | Files after | Time before | Time after |
| --- | --- | --- | --- | --- |
| 1 | 245 | 245 | 534 | 379 |
| 2 | 245 | 245 | 982 | 812 |

When the project is dirty (`mvn clean verify -DskipTests`), it is first imported (1), and then the workspace is re-opened (2).

| | Files before | Files after | Time before | Time after |
| --- | --- | --- | --- | --- |
| 1 | 246 | 245 | 520 | 376 |
| 2 | 246 | 245 | 1072 | 949 |